+++ core/arch/arm/mm/sub.mk
@@ -9,3 +9,4 @@ endif
 srcs-y += tee_mm.c
 srcs-y += pgt_cache.c
 srcs-y += mobj.c
+srcs-$(CFG_CORE_DYN_SHM) += mobj_dyn_shm.c
