+++ core/kernel/sub.mk
@@ -7,7 +7,7 @@ srcs-y += pm.c
 srcs-y += handle.c
 srcs-y += interrupt.c
 srcs-$(CFG_LOCKDEP) += lockdep.c
-srcs-y += msg_param.c
+srcs-$(CFG_CORE_DYN_SHM) += msg_param.c
 srcs-y += panic.c
 srcs-y += refcount.c
 srcs-y += tee_misc.c
