+++ mk/config.mk
@@ -402,12 +402,16 @@ CFG_DEVICE_ENUM_PTA ?= y
 # Default is 2**(2) = 4 cores per cluster.
 CFG_CORE_CLUSTER_SHIFT ?= 2
 
+# Enable support for dynamic shared memory (shared memory anywhere in
+# non-secure memory).
+CFG_CORE_DYN_SHM ?= y
+
 # Do not report to NW that dynamic shared memory (shared memory outside
 # predefined region) is enabled.
 # Note that you can disable this feature for debug purposes. OP-TEE will not
 # report to Normal World that it support dynamic SHM. But, nevertheles it
 # will accept dynamic SHM buffers.
-CFG_DYN_SHM_CAP ?= y
+CFG_DYN_SHM_CAP ?= $(CFG_CORE_DYN_SHM)
 
 # Enables support for larger physical addresses, that is, it will define
 # paddr_t as a 64-bit type.
